backend/pyproject.toml
[tool.poetry]
name = "firnow"
version = "0.1.0"
description = "Backend for FIRNow."
authors = ["Tsubasa <scientistjishan@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
aiohttp = "^3.8.5"
firebase-admin = "^6.2.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.0"}
pydantic = {extras = ["email"], version = "^2.4.2"}
fastapi = "^0.103.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
redis = {extras = ["hiredis"], version = "^5.0.1"}
pydantic-settings = "^2.0.3"
streaming-form-data = "^1.13.0"
aiofiles = "^23.2.1"
python-multipart = "^0.0.6"
w3storage = "^0.0.1"
typing-extensions = "^4.8.0"


[tool.poetry.group.dev.dependencies]
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
